# spring-boot-swagger/src/main/resources/static/swagger/swagger.yaml
{
    "swagger": "2.0",
    "info": {
        "description": "更多 Spring Boot 相关文章 https://microphoneben.github.io/",
        "version": "V1.0",
        "title": "Spring Boot 中基于Swagger2构建 RESTfulAPI 文档",
        "termsOfService": "http://www.baidu.com"
    },
    "host": "127.0.0.1:8080",
    "basePath": "/",
    "tags": [
        {
            "name": "Ben的Swagger测试",
            "description": "RESTful API"
        }
    ],
    "paths": {
        "/user": {
            "get": {
                "tags": [
                    "ben-controller"
                ],
                "summary": "获取用户列表",
                "operationId": "getUserListUsingGET",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "*/*"
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/User"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            },
            "post": {
                "tags": [
                    "ben-controller"
                ],
                "summary": "创建用户",
                "description": "根据User对象创建用户",
                "operationId": "postUserUsingPOST",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "*/*"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "description": "id",
                        "required": false,
                        "type": "integer",
                        "format": "int64"
                    },
                    {
                        "name": "name",
                        "in": "query",
                        "description": "姓名",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "age",
                        "in": "query",
                        "description": "年龄",
                        "required": false,
                        "type": "integer",
                        "format": "int32"
                    },
                    {
                        "in": "body",
                        "name": "user",
                        "description": "用户详细实体user",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/User"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "201": {
                        "description": "Created"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            }
        },
        "/user/{id}": {
            "get": {
                "tags": [
                    "ben-controller"
                ],
                "summary": "获取用户详细信息",
                "description": "根据url的id来获取用户详细信息",
                "operationId": "getUserUsingGET",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "*/*"
                ],
                "parameters": [
                    {
                        "in": "body",
                        "name": "id",
                        "description": "用户ID",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/Long"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/User"
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            },
            "put": {
                "tags": [
                    "ben-controller"
                ],
                "summary": "更新用户详细信息",
                "description": "根据url的id来指定更新对象，并根据传过来的user信息来更新用户详细信息",
                "operationId": "putUserUsingPUT",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "*/*"
                ],
                "parameters": [
                    {
                        "in": "body",
                        "name": "id",
                        "description": "用户ID",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/Long"
                        }
                    },
                    {
                        "in": "body",
                        "name": "user",
                        "description": "用户详细实体user",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/User"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "201": {
                        "description": "Created"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "ben-controller"
                ],
                "summary": "删除用户",
                "description": "根据url的id来指定删除对象",
                "operationId": "deleteUserUsingDELETE",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "*/*"
                ],
                "parameters": [
                    {
                        "in": "body",
                        "name": "id",
                        "description": "用户ID",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/Long"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "204": {
                        "description": "No Content"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        }
    },
    "definitions": {
        "User": {
            "type": "object",
            "properties": {
                "age": {
                    "type": "integer",
                    "format": "int32",
                    "description": "年龄"
                },
                "id": {
                    "type": "integer",
                    "format": "int64",
                    "description": "id"
                },
                "name": {
                    "type": "string",
                    "description": "姓名"
                }
            },
            "description": "请求入参 User"
        }
    }
}